Add tests for disaggregated Encoder. (#777)
* Add tests for disaggregated Encoder.


* Update DEVELOPMENT.md.


* Address review comments.


* Address review comments: add video_url tests.


* Address review comments: Add tests for audio_url and direct image embeddings Input.


* e2e_test.go: Extract helpers and configs into specific files.


* Simplify requests_test.go.


* Address review comments.


* Address review comments.


* Unify functions in setup_test.go and requests_test.go.


* Addess review comment: replace single-profile-handler with disagg-profile-handler.


* Minor function renaming.


* bump the simulator version to 0.8.2 (#73)


* Address review comments.


* Address review comments.


* Update image URLs.


* Address review comments.


---------
