+++ deploy/environments/dev/kind-istio-epd/kustomization.yaml
@@ -1,0 +1,17 @@
+# ------------------------------------------------------------------------------
+# Kubernetes In Docker (KIND) Environment
+#
+# This will deploy the full development stack on a KIND cluster:
+#
+#  * Istio Control Plane
+#  * VLLM Simulator
+#  * Inference Gateway
+# 
+# This will expose the VLLM simulator via InferencePool and an HTTPRoute.
+# ------------------------------------------------------------------------------
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+
+resources:
+- ../base-kind-istio/
+- ../../../components/vllm-sim-epd/
