+++ test/e2e/e2e_test.go
@@ -2,21 +2,11 @@ package e2e
 
 import (
 	"fmt"
-	"io"
-	"net/http"
-	"strconv"
-	"strings"
 	"time"
 
+	"github.com/llm-d/llm-d-inference-scheduler/pkg/metrics"
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
-	"github.com/openai/openai-go"
-	"github.com/openai/openai-go/option"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-	healthPb "google.golang.org/grpc/health/grpc_health_v1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	testutils "sigs.k8s.io/gateway-api-inference-extension/test/utils"
 )
 
@@ -24,22 +14,46 @@ const (
 	// simDeployment references the YAML file for the deployment
 	// running the vLLM simulator without PD
 	simDeployment = "./yaml/vllm-sim.yaml"
-	// simPDDeployment references the YAML file for the deployment
+	// simPDDisaggDeployment references the YAML file for the deployment
 	// running the vLLM simulator with PD (connector type is configurable via ${CONNECTOR_TYPE})
-	simPDDeployment = "./yaml/vllm-sim-pd.yaml"
+	simPDDisaggDeployment = "./yaml/vllm-sim-p-d-disagg.yaml"
 	// simDPDeployment references  the YAML file for the deployment
 	// running the vLLM simulator with Data Parallel
 	simDPDeployment = "./yaml/vllm-sim-dp.yaml"
+	// simEpDDisaggDeployment references the YAML file for the deployment
+	// running the vLLM simulator with E/PD (Encode/Prefill-Decode)
+	simEpDDisaggDeployment = "./yaml/vllm-sim-e-pd-disagg.yaml"
+	// simEPDDisaggDeployment references the YAML file for the deployment
+	// running the vLLM simulator with E/P/D (Encode/Prefill/Decode)
+	simEPDDisaggDeployment = "./yaml/vllm-sim-e-p-d-disagg.yaml"
+	// simEPDUnifiedDeployment references the YAML file for the deployment
+	// running the vLLM simulator with EPD (no disaggregation)
+	simEPDUnifiedDeployment = "./yaml/vllm-sim-epd-unified.yaml"
 
 	simplePrompt = "Hello my name is Andrew, I have a doctorate in Rocket Science, and I like interplanetary space exploration"
 	extraPrompt  = "Why is the sky sometimes blue and sometimes red close to sunset?"
+
+	// testImageURL and testImageURL2 are architecture diagrams stored in docs/images/ and served via GitHub raw content.
+	testImageURL  = "https://vllm-public-assets.s3.us-west-2.amazonaws.com/multimodal_asset/cat_snow.jpg"
+	testImageURL2 = "https://vllm-public-assets.s3.us-west-2.amazonaws.com/multimodal_asset/flycatcher.jpeg"
+	// testVideoURL is a publicly accessible video used in multimodal e2e tests.
+	testVideoURL = "https://www.bogotobogo.com/python/OpenCV_Python/images/mean_shift_tracking/slow_traffic_small.mp4"
+	// testImageEmbeds is a small dummy base64-encoded tensor used to test image_embeds requests.
+	// The actual bytes are not processed by the simulator; only routing behaviour is validated.
+	testImageEmbeds = "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=="
+	// testAudioData is a minimal base64-encoded WAV clip (44-byte header, no samples).
+	// The actual bytes are not processed by the simulator; only routing behaviour is validated.
+	testAudioData = "UklGRiQAAABXQVZFZm10IBAAAAABAAEAgD4AAAB9AAACABAAZGF0YQAAAAA="
 )
 
 var (
-	poolName        = simModelName + "-inference-pool"
-	podSelector     = map[string]string{"app": poolName}
-	prefillSelector = map[string]string{"llm-d.ai/role": "prefill"}
-	decodeSelector  = map[string]string{"llm-d.ai/role": "decode"}
+	poolName              = simModelName + "-inference-pool"
+	podSelector           = map[string]string{"app": poolName}
+	prefillSelector       = map[string]string{"llm-d.ai/role": "prefill"}
+	decodeSelector        = map[string]string{"llm-d.ai/role": "decode"}
+	prefillDecodeSelector = map[string]string{"llm-d.ai/role": "prefill-decode"}
+	encodeSelector        = map[string]string{"llm-d.ai/role": "encode"}
+	epdSingleSelector     = map[string]string{"llm-d.ai/role": "encode-prefill-decode"}
 )
 
 var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
@@ -47,7 +61,7 @@ var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
 		ginkgo.It("should run successfully", func() {
 			infPoolObjects = createInferencePool(1, true)
 
-			modelServers := createModelServers(false, false, false, 1, 0, 0)
+			modelServers := createModelServersDecode(1)
 
 			epp := createEndPointPicker(simpleConfig)
 
@@ -68,13 +82,258 @@ var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
 		})
 	})
 
-	ginkgo.When("Running a PD configuration", func() {
+	ginkgo.When("Running a PD configuration (deprecated pd-profile-handler)", func() {
 		ginkgo.It("should run successfully", func() {
 			infPoolObjects = createInferencePool(1, true)
 
 			prefillReplicas := 1
 			decodeReplicas := 4
-			modelServers := createModelServers(true, false, false, 0, prefillReplicas, decodeReplicas)
+			modelServers := createModelServersPDNixl(prefillReplicas, decodeReplicas)
+
+			epp := createEndPointPicker(deprecatedPdConfig)
+
+			metricsURL := fmt.Sprintf("http://localhost:%s/metrics", metricsPort)
+
+			if k8sContext != "" {
+				// Use port-forward to access the EPP pod's metrics endpoint.
+				startEPPMetricsPortForward()
+			}
+
+			prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
+			gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
+			gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
+
+			nsHdr, podHdrCompletion, _ := runCompletion(simplePrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdrCompletion).Should(gomega.BeElementOf(decodePods))
+
+			nsHdr, podHdrChat, _ := runChatCompletion(simplePrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdrChat).Should(gomega.BeElementOf(decodePods))
+
+			// Do an extra completion call with a different prompt
+			nsHdr, podHdr, _ := runCompletion(extraPrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+			// Run completion with the original prompt
+			nsHdr, podHdr, _ = runCompletion(simplePrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+			gomega.Expect(podHdr).Should(gomega.Equal(podHdrCompletion))
+
+			// Do an extra chat completion call with a different prompt
+			nsHdr, podHdr, _ = runChatCompletion(extraPrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+			// Run chat completion with the original prompt
+			nsHdr, podHdr, _ = runChatCompletion(simplePrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+			gomega.Expect(podHdr).Should(gomega.Equal(podHdrChat))
+
+			// Metrics Validation
+			labelFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypePrefillDecode, simModelName)
+			prefillDecodeCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_pd_decision_total", labelFilter)
+
+			labelFilter2 := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypeDecodeOnly, simModelName)
+			decodeOnlyCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_pd_decision_total", labelFilter2)
+
+			gomega.Expect(prefillDecodeCount).Should(gomega.Equal(4))
+			gomega.Expect(decodeOnlyCount).Should(gomega.Equal(2))
+
+			testutils.DeleteObjects(testConfig, epp)
+			testutils.DeleteObjects(testConfig, modelServers)
+		})
+	})
+
+	for _, tc := range []struct {
+		name   string
+		config string
+	}{
+		{"deprecated pd-profile-handler", deprecatedPdConfig},
+		{"disagg-profile-handler", pdConfig},
+	} {
+		config := tc.config // capture for closure
+		ginkgo.When("Running a PD configuration with shared-storage connector using "+tc.name, func() {
+			ginkgo.It("should run regular (non-streaming) requests successfully", func() {
+				infPoolObjects = createInferencePool(1, true)
+
+				prefillReplicas := 1
+				decodeReplicas := 2
+				modelServers := createModelServersPDSharedStorage(decodeReplicas)
+
+				epp := createEndPointPicker(config)
+
+				prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
+				gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
+				gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
+
+				// Test regular completion request
+				nsHdr, podHdrCompletion, _ := runCompletion(simplePrompt, simModelName)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdrCompletion).Should(gomega.BeElementOf(decodePods))
+
+				// Test regular chat completion request
+				nsHdr, podHdrChat, _ := runChatCompletion(simplePrompt, simModelName)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdrChat).Should(gomega.BeElementOf(decodePods))
+
+				// Run completion with a different prompt
+				nsHdr, podHdr, _ := runCompletion(extraPrompt, simModelName)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+				// Run completion with original prompt (should go to same pod due to prefix cache)
+				nsHdr, podHdr, _ = runCompletion(simplePrompt, simModelName)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+				gomega.Expect(podHdr).Should(gomega.Equal(podHdrCompletion))
+
+				testutils.DeleteObjects(testConfig, epp)
+				testutils.DeleteObjects(testConfig, modelServers)
+			})
+
+			ginkgo.It("should run streaming requests successfully", func() {
+				infPoolObjects = createInferencePool(1, true)
+
+				prefillReplicas := 1
+				decodeReplicas := 2
+				modelServers := createModelServersPDSharedStorage(decodeReplicas)
+
+				epp := createEndPointPicker(config)
+
+				prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
+				gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
+				gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
+
+				// Test streaming completion request
+				nsHdr, podHdr := runStreamingCompletion(simplePrompt, simModelName)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+				// Test streaming chat completion request
+				nsHdr, podHdr = runStreamingChatCompletion(simplePrompt)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+				// Run streaming completion with a different prompt
+				nsHdr, podHdr = runStreamingCompletion(extraPrompt, simModelName)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+				testutils.DeleteObjects(testConfig, epp)
+				testutils.DeleteObjects(testConfig, modelServers)
+			})
+
+			ginkgo.It("should handle decode-first success scenario with cache_hit_threshold", func() {
+				// This test verifies the decode-first optimization:
+				// When cache_hit_threshold is set and the decode succeeds (cache hit),
+				// the request should complete without falling back to P/D.
+				// IMPORTANT: The prefill pod should NOT process any requests in this scenario.
+				infPoolObjects = createInferencePool(1, true)
+
+				prefillReplicas := 1
+				decodeReplicas := 2
+				modelServers := createModelServersPDSharedStorage(decodeReplicas)
+
+				epp := createEndPointPicker(config)
+
+				prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
+				gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
+				gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
+
+				// Get prefill request count BEFORE the test
+				prefillCountBefore := getPodRequestCount(prefillPods[0])
+				ginkgo.By(fmt.Sprintf("Prefill request count before decode-first test: %d", prefillCountBefore))
+
+				// Test decode-first success: cache_hit_threshold is set, but simulator returns "stop"
+				// (without X-Cache-Threshold header), meaning decode succeeded without prefill
+				nsHdr, podHdr, finishReason := runCompletionWithCacheThreshold(simplePrompt, 0.5, false)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+				gomega.Expect(finishReason).ShouldNot(gomega.Equal("cache_threshold"))
+
+				// Test streaming decode-first success
+				nsHdr, podHdr, finishReason = runStreamingCompletionWithCacheThreshold(simplePrompt, 0.5, false)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+				gomega.Expect(finishReason).ShouldNot(gomega.Equal("cache_threshold"))
+
+				// Get prefill request count AFTER the test
+				prefillCountAfter := getPodRequestCount(prefillPods[0])
+				ginkgo.By(fmt.Sprintf("Prefill request count after decode-first test: %d", prefillCountAfter))
+
+				// VERIFY: Prefill pod should NOT have processed any new requests
+				// (decode-first succeeded, so no P/D fallback occurred)
+				gomega.Expect(prefillCountAfter).Should(gomega.Equal(prefillCountBefore),
+					"Prefill pod should NOT process requests when cache threshold is met (decode-first success)")
+
+				testutils.DeleteObjects(testConfig, epp)
+				testutils.DeleteObjects(testConfig, modelServers)
+			})
+
+			ginkgo.It("should handle decode-first fallback to P/D when cache threshold not met", func() {
+				// This test verifies the decode-first fallback scenario:
+				// When cache_hit_threshold is set and the decode returns cache_threshold finish_reason,
+				// the sidecar should fall back to P/D disaggregation.
+				// IMPORTANT: The prefill pod SHOULD process requests in this scenario.
+				infPoolObjects = createInferencePool(1, true)
+
+				prefillReplicas := 1
+				decodeReplicas := 2
+				modelServers := createModelServersPDSharedStorage(decodeReplicas)
+
+				epp := createEndPointPicker(config)
+
+				prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
+				gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
+				gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
+
+				// Get prefill request count BEFORE the test
+				prefillCountBefore := getPodRequestCount(prefillPods[0])
+				ginkgo.By(fmt.Sprintf("Prefill request count before P/D fallback test: %d", prefillCountBefore))
+
+				// Test decode-first fallback: cache_hit_threshold is set AND X-Cache-Threshold header
+				// forces simulator to return "cache_threshold" finish_reason, triggering P/D fallback
+				nsHdr, podHdr, finishReason := runCompletionWithCacheThreshold(simplePrompt, 0.5, true)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+				// The sidecar completes the P/D flow but returns cache_threshold as the finish_reason
+				// from the initial decode attempt (which triggered the fallback)
+				gomega.Expect(finishReason).Should(gomega.Equal("cache_threshold"))
+
+				// Test streaming decode-first fallback
+				nsHdr, podHdr, finishReason = runStreamingCompletionWithCacheThreshold(extraPrompt, 0.5, true)
+				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+				gomega.Expect(finishReason).Should(gomega.Equal("cache_threshold"))
+
+				// Get prefill request count AFTER the test
+				prefillCountAfter := getPodRequestCount(prefillPods[0])
+				ginkgo.By(fmt.Sprintf("Prefill request count after P/D fallback test: %d", prefillCountAfter))
+
+				// VERIFY: Prefill pod SHOULD have processed 2 new requests (1 regular + 1 streaming)
+				// (decode-first failed, so P/D fallback occurred and prefill was invoked)
+				gomega.Expect(prefillCountAfter).Should(gomega.BeNumerically(">", prefillCountBefore),
+					"Prefill pod SHOULD process requests when cache threshold is NOT met (P/D fallback)")
+				gomega.Expect(prefillCountAfter-prefillCountBefore).Should(gomega.Equal(2),
+					"Prefill pod should have processed exactly 2 requests (1 regular + 1 streaming)")
+
+				testutils.DeleteObjects(testConfig, epp)
+				testutils.DeleteObjects(testConfig, modelServers)
+			})
+		})
+	}
+
+	ginkgo.When("Running a PD configuration with disagg-profile-handler and metrics validation", func() {
+		ginkgo.It("should run successfully", func() {
+			infPoolObjects = createInferencePool(1, true)
+
+			prefillReplicas := 1
+			decodeReplicas := 4
+			modelServers := createModelServersPDSharedStorage(decodeReplicas)
 
 			epp := createEndPointPicker(pdConfig)
 
@@ -120,255 +379,10 @@ var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
 			gomega.Expect(podHdr).Should(gomega.Equal(podHdrChat))
 
 			// Metrics Validation
-			labelFilter := fmt.Sprintf(`decision_type="prefill-decode",model_name="%s"`, simModelName)
-			prefillDecodeCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_pd_decision_total", labelFilter)
-
-			labelFilter2 := fmt.Sprintf(`decision_type="decode-only",model_name="%s"`, simModelName)
-			decodeOnlyCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_pd_decision_total", labelFilter2)
-
-			gomega.Expect(prefillDecodeCount).Should(gomega.Equal(4))
-			gomega.Expect(decodeOnlyCount).Should(gomega.Equal(2))
-
-			testutils.DeleteObjects(testConfig, epp)
-			testutils.DeleteObjects(testConfig, modelServers)
-		})
-	})
-
-	for _, tc := range []struct {
-		name   string
-		config string
-	}{
-		{"pd-profile-handler", pdConfig},
-		{"disagg-profile-handler", pdDisaggConfig},
-	} {
-		config := tc.config // capture for closure
-		ginkgo.When("Running a PD configuration with shared-storage connector using "+tc.name, func() {
-			ginkgo.It("should run regular (non-streaming) requests successfully", func() {
-				infPoolObjects = createInferencePool(1, true)
-
-				prefillReplicas := 1
-				decodeReplicas := 2
-				modelServers := createModelServersWithConnector(true, false, false, 0, prefillReplicas, decodeReplicas, "shared-storage")
-
-				epp := createEndPointPicker(config)
-
-				prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
-				gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
-				gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
-
-				// Test regular completion request
-				nsHdr, podHdrCompletion, _ := runCompletion(simplePrompt, simModelName)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdrCompletion).Should(gomega.BeElementOf(decodePods))
-
-				// Test regular chat completion request
-				nsHdr, podHdrChat, _ := runChatCompletion(simplePrompt, simModelName)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdrChat).Should(gomega.BeElementOf(decodePods))
-
-				// Run completion with a different prompt
-				nsHdr, podHdr, _ := runCompletion(extraPrompt, simModelName)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-
-				// Run completion with original prompt (should go to same pod due to prefix cache)
-				nsHdr, podHdr, _ = runCompletion(simplePrompt, simModelName)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-				gomega.Expect(podHdr).Should(gomega.Equal(podHdrCompletion))
-
-				testutils.DeleteObjects(testConfig, epp)
-				testutils.DeleteObjects(testConfig, modelServers)
-			})
-
-			ginkgo.It("should run streaming requests successfully", func() {
-				infPoolObjects = createInferencePool(1, true)
-
-				prefillReplicas := 1
-				decodeReplicas := 2
-				modelServers := createModelServersWithConnector(true, false, false, 0, prefillReplicas, decodeReplicas, "shared-storage")
-
-				epp := createEndPointPicker(config)
-
-				prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
-				gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
-				gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
-
-				// Test streaming completion request
-				nsHdr, podHdr := runStreamingCompletion(simplePrompt, simModelName)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-
-				// Test streaming chat completion request
-				nsHdr, podHdr = runStreamingChatCompletion(simplePrompt)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-
-				// Run streaming completion with a different prompt
-				nsHdr, podHdr = runStreamingCompletion(extraPrompt, simModelName)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-
-				testutils.DeleteObjects(testConfig, epp)
-				testutils.DeleteObjects(testConfig, modelServers)
-			})
-
-			ginkgo.It("should handle decode-first success scenario with cache_hit_threshold", func() {
-				// This test verifies the decode-first optimization:
-				// When cache_hit_threshold is set and the decode succeeds (cache hit),
-				// the request should complete without falling back to P/D.
-				// IMPORTANT: The prefill pod should NOT process any requests in this scenario.
-				infPoolObjects = createInferencePool(1, true)
-
-				prefillReplicas := 1
-				decodeReplicas := 2
-				modelServers := createModelServersWithConnector(true, false, false, 0, prefillReplicas, decodeReplicas, "shared-storage")
-
-				epp := createEndPointPicker(config)
-
-				prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
-				gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
-				gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
-
-				// Get prefill request count BEFORE the test
-				prefillCountBefore := getPrefillRequestCount(prefillPods[0])
-				ginkgo.By(fmt.Sprintf("Prefill request count before decode-first test: %d", prefillCountBefore))
-
-				// Test decode-first success: cache_hit_threshold is set, but simulator returns "stop"
-				// (without X-Cache-Threshold header), meaning decode succeeded without prefill
-				nsHdr, podHdr, finishReason := runCompletionWithCacheThreshold(simplePrompt, 0.5, false)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-				gomega.Expect(finishReason).ShouldNot(gomega.Equal("cache_threshold"))
-
-				// Test streaming decode-first success
-				nsHdr, podHdr, finishReason = runStreamingCompletionWithCacheThreshold(simplePrompt, 0.5, false)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-				gomega.Expect(finishReason).ShouldNot(gomega.Equal("cache_threshold"))
-
-				// Get prefill request count AFTER the test
-				prefillCountAfter := getPrefillRequestCount(prefillPods[0])
-				ginkgo.By(fmt.Sprintf("Prefill request count after decode-first test: %d", prefillCountAfter))
-
-				// VERIFY: Prefill pod should NOT have processed any new requests
-				// (decode-first succeeded, so no P/D fallback occurred)
-				gomega.Expect(prefillCountAfter).Should(gomega.Equal(prefillCountBefore),
-					"Prefill pod should NOT process requests when cache threshold is met (decode-first success)")
-
-				testutils.DeleteObjects(testConfig, epp)
-				testutils.DeleteObjects(testConfig, modelServers)
-			})
-
-			ginkgo.It("should handle decode-first fallback to P/D when cache threshold not met", func() {
-				// This test verifies the decode-first fallback scenario:
-				// When cache_hit_threshold is set and the decode returns cache_threshold finish_reason,
-				// the sidecar should fall back to P/D disaggregation.
-				// IMPORTANT: The prefill pod SHOULD process requests in this scenario.
-				infPoolObjects = createInferencePool(1, true)
-
-				prefillReplicas := 1
-				decodeReplicas := 2
-				modelServers := createModelServersWithConnector(true, false, false, 0, prefillReplicas, decodeReplicas, "shared-storage")
-
-				epp := createEndPointPicker(config)
-
-				prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
-				gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
-				gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
-
-				// Get prefill request count BEFORE the test
-				prefillCountBefore := getPrefillRequestCount(prefillPods[0])
-				ginkgo.By(fmt.Sprintf("Prefill request count before P/D fallback test: %d", prefillCountBefore))
-
-				// Test decode-first fallback: cache_hit_threshold is set AND X-Cache-Threshold header
-				// forces simulator to return "cache_threshold" finish_reason, triggering P/D fallback
-				nsHdr, podHdr, finishReason := runCompletionWithCacheThreshold(simplePrompt, 0.5, true)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-				// The sidecar completes the P/D flow but returns cache_threshold as the finish_reason
-				// from the initial decode attempt (which triggered the fallback)
-				gomega.Expect(finishReason).Should(gomega.Equal("cache_threshold"))
-
-				// Test streaming decode-first fallback
-				nsHdr, podHdr, finishReason = runStreamingCompletionWithCacheThreshold(extraPrompt, 0.5, true)
-				gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-				gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-				gomega.Expect(finishReason).Should(gomega.Equal("cache_threshold"))
-
-				// Get prefill request count AFTER the test
-				prefillCountAfter := getPrefillRequestCount(prefillPods[0])
-				ginkgo.By(fmt.Sprintf("Prefill request count after P/D fallback test: %d", prefillCountAfter))
-
-				// VERIFY: Prefill pod SHOULD have processed 2 new requests (1 regular + 1 streaming)
-				// (decode-first failed, so P/D fallback occurred and prefill was invoked)
-				gomega.Expect(prefillCountAfter).Should(gomega.BeNumerically(">", prefillCountBefore),
-					"Prefill pod SHOULD process requests when cache threshold is NOT met (P/D fallback)")
-				gomega.Expect(prefillCountAfter-prefillCountBefore).Should(gomega.Equal(2),
-					"Prefill pod should have processed exactly 2 requests (1 regular + 1 streaming)")
-
-				testutils.DeleteObjects(testConfig, epp)
-				testutils.DeleteObjects(testConfig, modelServers)
-			})
-		})
-	}
-
-	ginkgo.When("Running a PD configuration with disagg-profile-handler and metrics validation", func() {
-		ginkgo.It("should run successfully", func() {
-			infPoolObjects = createInferencePool(1, true)
-
-			prefillReplicas := 1
-			decodeReplicas := 4
-			modelServers := createModelServersWithConnector(true, false, false, 0, prefillReplicas, decodeReplicas, "shared-storage")
-
-			epp := createEndPointPicker(pdDisaggConfig)
-
-			metricsURL := fmt.Sprintf("http://localhost:%s/metrics", metricsPort)
-
-			if k8sContext != "" {
-				// Use port-forward to access the EPP pod's metrics endpoint.
-				startEPPMetricsPortForward()
-			}
-
-			prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
-			gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
-			gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
-
-			nsHdr, podHdrCompletion, _ := runCompletion(simplePrompt, simModelName)
-			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-			gomega.Expect(podHdrCompletion).Should(gomega.BeElementOf(decodePods))
-
-			nsHdr, podHdrChat, _ := runChatCompletion(simplePrompt, simModelName)
-			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-			gomega.Expect(podHdrChat).Should(gomega.BeElementOf(decodePods))
-
-			// Do an extra completion call with a different prompt
-			nsHdr, podHdr, _ := runCompletion(extraPrompt, simModelName)
-			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-
-			// Run completion with the original prompt
-			nsHdr, podHdr, _ = runCompletion(simplePrompt, simModelName)
-			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-			gomega.Expect(podHdr).Should(gomega.Equal(podHdrCompletion))
-
-			// Do an extra chat completion call with a different prompt
-			nsHdr, podHdr, _ = runChatCompletion(extraPrompt, simModelName)
-			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-
-			// Run chat completion with the original prompt
-			nsHdr, podHdr, _ = runChatCompletion(simplePrompt, simModelName)
-			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
-			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
-			gomega.Expect(podHdr).Should(gomega.Equal(podHdrChat))
-
-			// Metrics Validation
-			labelFilter := fmt.Sprintf(`decision_type="prefill-decode",model_name="%s"`, simModelName)
+			labelFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypePrefillDecode, simModelName)
 			prefillDecodeCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", labelFilter)
 
-			labelFilter2 := fmt.Sprintf(`decision_type="decode-only",model_name="%s"`, simModelName)
+			labelFilter2 := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypeDecodeOnly, simModelName)
 			decodeOnlyCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", labelFilter2)
 
 			gomega.Expect(prefillDecodeCount).Should(gomega.Equal(4))
@@ -383,9 +397,9 @@ var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
 		ginkgo.It("should run successfully", func() {
 			infPoolObjects = createInferencePool(1, true)
 
-			modelServers := createModelServers(false, false, false, 1, 0, 0)
+			modelServers := createModelServersDecode(1)
 
-			epp := createEndPointPicker(disaggDecodeOnlyConfig)
+			epp := createEndPointPicker(decodeOnlyConfig)
 
 			prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
 			gomega.Expect(prefillPods).Should(gomega.BeEmpty())
@@ -404,13 +418,220 @@ var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
 		})
 	})
 
+	ginkgo.When("Running an E/PD (Encode/Prefill-Decode) configuration", func() {
+		ginkgo.It("should route multimodal requests through encode and decode pods", func() {
+			infPoolObjects = createInferencePool(1, true)
+
+			encodeReplicas := 2
+			decodeReplicas := 1
+			modelServers := createModelServersEpDDisagg(encodeReplicas, decodeReplicas)
+
+			epp := createEndPointPicker(epdEncodeDecodeConfig)
+
+			metricsURL := fmt.Sprintf("http://localhost:%s/metrics", metricsPort)
+			if k8sContext != "" {
+				startEPPMetricsPortForward()
+			}
+
+			encodePods := getPodNames(encodeSelector)
+			prefillDecodePods := getPodNames(prefillDecodeSelector)
+			gomega.Expect(encodePods).Should(gomega.HaveLen(encodeReplicas))
+			gomega.Expect(prefillDecodePods).Should(gomega.HaveLen(decodeReplicas))
+
+			// Text request: encode stage skipped, routed directly to a prefill-decode pod
+			nsHdr, podHdr, _ := runCompletion(simplePrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(prefillDecodePods))
+
+			// Multimodal request: triggers encode stage, decode handled by prefill-decode pod
+			nsHdr, podHdr = runChatCompletionWithImages(testImageURL)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(prefillDecodePods))
+
+			nsHdr, podHdr = runChatCompletionWithImages(testImageURL)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(prefillDecodePods))
+
+			// Multi-image request: two images in one request, triggers encode stage
+			nsHdr, podHdr = runChatCompletionWithImages(testImageURL, testImageURL2)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(prefillDecodePods))
+
+			// Video request: video_url triggers encode stage, decode handled by prefill-decode pod
+			nsHdr, podHdr = runChatCompletionWithVideo()
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(prefillDecodePods))
+
+			// Audio request: input_audio triggers encode stage, decode handled by prefill-decode pod
+			nsHdr, podHdr = runChatCompletionWithAudio()
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(prefillDecodePods))
+
+			// image_embeds request: pre-encoded tensor, encode stage skipped, routes to prefill-decode pod
+			nsHdr, podHdr = runChatCompletionWithImageEmbeds()
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(prefillDecodePods))
+
+			// Metrics: text + image_embeds requests recorded as decode-only (encode skipped)
+			decodeOnlyFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypeDecodeOnly, simModelName)
+			decodeOnlyCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", decodeOnlyFilter)
+			gomega.Expect(decodeOnlyCount).Should(gomega.Equal(2))
+
+			// Metrics: encode-decode decisions recorded (2 single-image + 1 multi-image + 1 video + 1 audio)
+			labelFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypeEncodeDecode, simModelName)
+			encodeDecodeCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", labelFilter)
+			gomega.Expect(encodeDecodeCount).Should(gomega.Equal(5))
+
+			testutils.DeleteObjects(testConfig, epp)
+			testutils.DeleteObjects(testConfig, modelServers)
+		})
+	})
+
+	ginkgo.When("Running an E/P/D (encode/prefill/decode) configuration", func() {
+		ginkgo.It("should route multimodal requests through encode, prefill, and decode pods", func() {
+			infPoolObjects = createInferencePool(1, true)
+
+			encodeReplicas := 2
+			prefillReplicas := 1
+			decodeReplicas := 1
+			modelServers := createModelServersEPDDisagg(encodeReplicas, prefillReplicas, decodeReplicas)
+
+			epp := createEndPointPicker(epdConfig)
+
+			metricsURL := fmt.Sprintf("http://localhost:%s/metrics", metricsPort)
+			if k8sContext != "" {
+				startEPPMetricsPortForward()
+			}
+
+			encodePods := getPodNames(encodeSelector)
+			prefillPods := getPodNames(prefillSelector)
+			decodePods := getPodNames(decodeSelector)
+			gomega.Expect(encodePods).Should(gomega.HaveLen(encodeReplicas))
+			gomega.Expect(prefillPods).Should(gomega.HaveLen(prefillReplicas))
+			gomega.Expect(decodePods).Should(gomega.HaveLen(decodeReplicas))
+
+			// Text request: encode stage skipped, prefill triggered by prefix-based-pd-decider
+			nsHdr, podHdr, _ := runCompletion(simplePrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+			// First multimodal request: encode + prefill + decode
+			nsHdr, podHdr = runChatCompletionWithImages(testImageURL)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+			// Second multimodal request with same image (prefix cache may skip prefill)
+			nsHdr, podHdr = runChatCompletionWithImages(testImageURL)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+			// Multi-image request: two images in one request, encode + prefill + decode
+			nsHdr, podHdr = runChatCompletionWithImages(testImageURL, testImageURL2)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+			// Video request: video_url triggers encode stage, decode handled by decode pod
+			nsHdr, podHdr = runChatCompletionWithVideo()
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+			// image_embeds request: pre-encoded tensor, encode stage skipped, routes to decode pod
+			nsHdr, podHdr = runChatCompletionWithImageEmbeds()
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.BeElementOf(decodePods))
+
+			// Metrics: text + image_embeds requests recorded as decode-only or prefill-decode (encode skipped)
+			pdLabelFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypePrefillDecode, simModelName)
+			doLabelFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypeDecodeOnly, simModelName)
+			pdCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", pdLabelFilter)
+			doCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", doLabelFilter)
+			gomega.Expect(pdCount + doCount).Should(gomega.Equal(2))
+
+			// Metrics: 4 multimodal requests each produce either encode-prefill-decode or encode-decode
+			// (encode-decode occurs if the prefix cache hits on the second same-image request).
+			// The 3 requests with unique content (1st image, multi-image, video) always produce encode-prefill-decode.
+			epdLabelFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypeEncodePrefillDecode, simModelName)
+			edLabelFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypeEncodeDecode, simModelName)
+			epdCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", epdLabelFilter)
+			edCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", edLabelFilter)
+			gomega.Expect(epdCount).Should(gomega.BeNumerically(">=", 3))
+			gomega.Expect(epdCount + edCount).Should(gomega.Equal(4))
+
+			testutils.DeleteObjects(testConfig, epp)
+			testutils.DeleteObjects(testConfig, modelServers)
+		})
+	})
+
+	ginkgo.When("Running an EPD (no disaggregation) configuration", func() {
+		ginkgo.It("should route text and multimodal requests to the single deployment", func() {
+			infPoolObjects = createInferencePool(1, true)
+
+			// Single deployment labeled encode-prefill-decode: matches encode-filter, prefill-filter,
+			// and decode-filter, so all EPD stages are handled by the same deployment.
+			replicas := 1
+			modelServers := createModelServersEPDUnified(replicas)
+
+			// Using epdConfig instead of decodeOnlyConfig to validate the EPD logic path within
+			// a single pod; multimodal stages will resolve to this same deployment.
+			epp := createEndPointPicker(epdConfig)
+
+			metricsURL := fmt.Sprintf("http://localhost:%s/metrics", metricsPort)
+			if k8sContext != "" {
+				startEPPMetricsPortForward()
+			}
+
+			epdPods := getPodNames(epdSingleSelector)
+			gomega.Expect(epdPods).Should(gomega.HaveLen(replicas))
+
+			// Text completion: encode skipped, routes to decode profile -> single deployment
+			nsHdr, podHdr, _ := runCompletion(simplePrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.Equal(epdPods[0]))
+
+			// Text chat completion: same routing as above
+			nsHdr, podHdr, _ = runChatCompletion(simplePrompt, simModelName)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.Equal(epdPods[0]))
+
+			// Metrics: text requests recorded as decode-only or prefill-decode (encode skipped)
+			pdLabelFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypePrefillDecode, simModelName)
+			doLabelFilter := fmt.Sprintf(`decision_type=%q,model_name="%s"`, metrics.DecisionTypeDecodeOnly, simModelName)
+			pdCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", pdLabelFilter)
+			doCount := getCounterMetric(metricsURL, "llm_d_inference_scheduler_disagg_decision_total", doLabelFilter)
+			gomega.Expect(pdCount + doCount).Should(gomega.Equal(2))
+
+			// Multimodal request: encode and decode profiles both resolve to the same single deployment
+			nsHdr, podHdr = runChatCompletionWithImages(testImageURL)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.Equal(epdPods[0]))
+
+			// Multi-image request: all stages handled by single deployment
+			nsHdr, podHdr = runChatCompletionWithImages(testImageURL, testImageURL2)
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.Equal(epdPods[0]))
+
+			// Video request: all stages handled by single deployment
+			nsHdr, podHdr = runChatCompletionWithVideo()
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.Equal(epdPods[0]))
+
+			// image_embeds request: encode skipped, routes to single deployment
+			nsHdr, podHdr = runChatCompletionWithImageEmbeds()
+			gomega.Expect(nsHdr).Should(gomega.Equal(nsName))
+			gomega.Expect(podHdr).Should(gomega.Equal(epdPods[0]))
+
+			testutils.DeleteObjects(testConfig, epp)
+			testutils.DeleteObjects(testConfig, modelServers)
+		})
+	})
+
 	ginkgo.When("Running simple non-PD KV enabled configuration", func() {
 		ginkgo.It("should run successfully", func() {
 			infPoolObjects = createInferencePool(1, true)
 
 			epp := createEndPointPicker(kvConfig)
 
-			modelServers := createModelServers(false, true, false, 1, 0, 0)
+			modelServers := createModelServersDecodeKV(1)
 			time.Sleep(5 * time.Second) // wait for model server(s) to become ready
 
 			prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
@@ -434,7 +655,7 @@ var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
 
 			epp := createEndPointPicker(kvExternalTokenizerConfig)
 
-			modelServers := createModelServers(false, true, false, 1, 0, 0)
+			modelServers := createModelServersDecodeKV(1)
 			time.Sleep(5 * time.Second) // wait for model server(s) to become ready
 
 			prefillPods, decodePods := getModelServerPods(podSelector, prefillSelector, decodeSelector)
@@ -467,7 +688,7 @@ var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
 		ginkgo.It("should distribute inference requests across all model servers", func() {
 			infPoolObjects = createInferencePool(1, true)
 
-			modelServers := createModelServers(false, false, false, 1, 0, 0)
+			modelServers := createModelServersDecode(1)
 
 			epp := createEndPointPicker(scaleConfig)
 
@@ -523,7 +744,7 @@ var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
 		ginkgo.It("should schedule inference on all ranks", func() {
 			infPoolObjects = createInferencePool(2, true)
 
-			modelServers := createModelServers(false, false, true, 1, 0, 0)
+			modelServers := createModelServersDecodeDP(1)
 
 			epp := createEndPointPicker(dataParallelConfig)
 
@@ -568,606 +789,3 @@ var _ = ginkgo.Describe("Run end to end tests", ginkgo.Ordered, func() {
 		})
 	})
 })
-
-// createModelServers creates the model server resources used for testing from the given filePaths.
-// Uses the default connector (nixlv2) for P/D deployments.
-func createModelServers(withPD, withKV, withDP bool, vllmReplicas, prefillReplicas, decodeReplicas int) []string {
-	return createModelServersWithConnector(withPD, withKV, withDP, vllmReplicas, prefillReplicas, decodeReplicas, "nixlv2")
-}
-
-// createModelServersWithConnector creates model server resources with a specific connector type.
-func createModelServersWithConnector(withPD, withKV, withDP bool, vllmReplicas, prefillReplicas, decodeReplicas int, connector string) []string {
-	theModelName := simModelName
-	theSafeModelName := simModelName
-	if withKV {
-		theModelName = kvModelName
-		theSafeModelName = safeKvModelName
-	}
-	yaml := simDeployment
-	if withPD {
-		yaml = simPDDeployment
-	} else if withDP {
-		yaml = simDPDeployment
-	}
-
-	manifests := testutils.ReadYaml(yaml)
-	manifests = substituteMany(manifests,
-		map[string]string{
-			"${MODEL_NAME}":           theModelName,
-			"${MODEL_NAME_SAFE}":      theSafeModelName,
-			"${POOL_NAME}":            poolName,
-			"${KV_CACHE_ENABLED}":     strconv.FormatBool(withKV),
-			"${CONNECTOR_TYPE}":       connector,
-			"${SIDECAR_IMAGE}":        sideCarImage,
-			"${VLLM_REPLICA_COUNT}":   strconv.Itoa(vllmReplicas),
-			"${VLLM_REPLICA_COUNT_D}": strconv.Itoa(decodeReplicas),
-			"${VLLM_REPLICA_COUNT_P}": strconv.Itoa(prefillReplicas),
-			"${VLLM_SIMULATOR_IMAGE}": vllmSimImage,
-			"${UDS_TOKENIZER_IMAGE}":  udsTokenizerImage,
-		})
-
-	objects := testutils.CreateObjsFromYaml(testConfig, manifests)
-	podsInDeploymentsReady(objects)
-
-	return objects
-}
-
-func createEndPointPicker(eppConfig string) []string {
-	configMap := &corev1.ConfigMap{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "v1",
-			Kind:       "ConfigMap",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "epp-config",
-			Namespace: nsName,
-		},
-		Data: map[string]string{"epp-config.yaml": eppConfig},
-	}
-	err := testConfig.K8sClient.Create(testConfig.Context, configMap)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-
-	objects := make([]string, 1, 10)
-	objects[0] = "ConfigMap/epp-config"
-
-	eppYamls := testutils.ReadYaml(eppManifest)
-	eppYamls = substituteMany(eppYamls,
-		map[string]string{
-			"${EPP_IMAGE}":           eppImage,
-			"${UDS_TOKENIZER_IMAGE}": udsTokenizerImage,
-			"${NAMESPACE}":           nsName,
-			"${POOL_NAME}":           simModelName + "-inference-pool",
-		})
-
-	objects = append(objects, testutils.CreateObjsFromYaml(testConfig, eppYamls)...)
-	podsInDeploymentsReady(objects)
-
-	ginkgo.By("Waiting for EPP to report that it is serving")
-	conn, err := grpc.NewClient("localhost:30081",
-		grpc.WithTransportCredentials(insecure.NewCredentials()))
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	defer func() {
-		err := conn.Close()
-		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	}()
-	client := healthPb.NewHealthClient(conn)
-	healthCheckReq := &healthPb.HealthCheckRequest{}
-
-	gomega.Eventually(func() bool {
-		resp, err := client.Check(testConfig.Context, healthCheckReq)
-		return err == nil && resp.Status == healthPb.HealthCheckResponse_SERVING
-	}, 40*time.Second, 2*time.Second).Should(gomega.BeTrue())
-	ginkgo.By("EPP reports that it is serving")
-	time.Sleep(2 * time.Second)
-
-	return objects
-}
-
-func runCompletion(prompt string, theModel openai.CompletionNewParamsModel) (string, string, string) {
-	var httpResp *http.Response
-	openaiclient := openai.NewClient(
-		option.WithBaseURL(fmt.Sprintf("http://localhost:%s/v1", port)))
-
-	completionParams := openai.CompletionNewParams{
-		Prompt: openai.CompletionNewParamsPromptUnion{
-			OfString: openai.String(prompt),
-		},
-		Model: theModel,
-	}
-
-	ginkgo.By(fmt.Sprintf("Sending Completion Request: (port %s) %#v", port, completionParams))
-
-	resp, err := openaiclient.Completions.New(testConfig.Context, completionParams, option.WithResponseInto(&httpResp), option.WithRequestTimeout(readyTimeout))
-
-	ginkgo.By(fmt.Sprintf("Verifying Completion Response: %#v", resp))
-
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	gomega.Expect(resp.Choices).Should(gomega.HaveLen(1))
-	gomega.Expect(resp.Choices[0].FinishReason).Should(gomega.Equal(openai.CompletionChoiceFinishReasonStop))
-	gomega.Expect(resp.Choices[0].Text).Should(gomega.Equal(prompt))
-
-	namespaceHeader := httpResp.Header.Get("x-inference-namespace")
-	podHeader := httpResp.Header.Get("x-inference-pod")
-	podPort := httpResp.Header.Get("x-inference-port")
-
-	return namespaceHeader, podHeader, podPort
-}
-
-func runChatCompletion(prompt, modelName string) (string, string, string) {
-	var httpResp *http.Response
-	openaiclient := openai.NewClient(
-		option.WithBaseURL(fmt.Sprintf("http://localhost:%s/v1", port)))
-
-	params := openai.ChatCompletionNewParams{
-		Messages: []openai.ChatCompletionMessageParamUnion{
-			openai.UserMessage(prompt),
-		},
-		Model: modelName,
-	}
-	resp, err := openaiclient.Chat.Completions.New(testConfig.Context, params, option.WithResponseInto(&httpResp))
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	gomega.Expect(resp.Choices).Should(gomega.HaveLen(1))
-	gomega.Expect(resp.Choices[0].FinishReason).Should(gomega.Equal("stop"))
-	gomega.Expect(resp.Choices[0].Message.Content).Should(gomega.Equal(prompt))
-
-	namespaceHeader := httpResp.Header.Get("x-inference-namespace")
-	podHeader := httpResp.Header.Get("x-inference-pod")
-	podPort := httpResp.Header.Get("x-inference-port")
-
-	return namespaceHeader, podHeader, podPort
-}
-
-// getCounterMetric fetches the current value of a Prometheus counter metric from the given metrics URL.
-//
-//nolint:unparam // metricName may vary in future test cases
-func getCounterMetric(metricsURL, metricName, labelMatch string) int {
-	resp, err := http.Get(metricsURL)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	defer func() {
-		err = resp.Body.Close()
-		gomega.Expect(err).ToNot(gomega.HaveOccurred())
-	}()
-	gomega.Expect(resp.StatusCode).Should(gomega.Equal(http.StatusOK))
-
-	body, err := io.ReadAll(resp.Body)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-
-	metricsText := string(body)
-	for _, line := range strings.Split(metricsText, "\n") {
-		if strings.HasPrefix(line, metricName) && strings.Contains(line, labelMatch) {
-			fields := strings.Fields(line)
-			if len(fields) >= 2 {
-				valFloat, err := strconv.ParseFloat(fields[len(fields)-1], 64)
-				gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-				return int(valFloat)
-			}
-		}
-	}
-	return 0
-}
-
-func runStreamingCompletion(prompt string, theModel openai.CompletionNewParamsModel) (string, string) {
-	ginkgo.By(fmt.Sprintf("Sending Streaming Completion Request: (port %s) model=%s", port, theModel))
-
-	// Use raw HTTP for streaming to capture headers
-	body := fmt.Sprintf(`{"model":"%s","prompt":"%s","max_tokens":50,"stream":true}`, theModel, prompt)
-	req, err := http.NewRequest(http.MethodPost, fmt.Sprintf("http://localhost:%s/v1/completions", port), strings.NewReader(body))
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := http.DefaultClient.Do(req)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	defer func() {
-		err := resp.Body.Close()
-		gomega.Expect(err).ToNot(gomega.HaveOccurred())
-	}()
-
-	gomega.Expect(resp.StatusCode).Should(gomega.Equal(http.StatusOK))
-
-	namespaceHeader := resp.Header.Get("x-inference-namespace")
-	podHeader := resp.Header.Get("x-inference-pod")
-
-	// Read and verify the streaming response
-	respBody, err := io.ReadAll(resp.Body)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-
-	ginkgo.By(fmt.Sprintf("Streaming Completion received response length: %d bytes", len(respBody)))
-
-	return namespaceHeader, podHeader
-}
-
-func runStreamingChatCompletion(prompt string) (string, string) {
-	ginkgo.By(fmt.Sprintf("Sending Streaming Chat Completion Request: (port %s)", port))
-
-	// Use raw HTTP for streaming to capture headers
-	body := fmt.Sprintf(`{"model":"%s","messages":[{"role":"user","content":"%s"}],"stream":true}`, simModelName, prompt)
-	req, err := http.NewRequest(http.MethodPost, fmt.Sprintf("http://localhost:%s/v1/chat/completions", port), strings.NewReader(body))
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := http.DefaultClient.Do(req)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	defer func() {
-		err := resp.Body.Close()
-		gomega.Expect(err).ToNot(gomega.HaveOccurred())
-	}()
-
-	gomega.Expect(resp.StatusCode).Should(gomega.Equal(http.StatusOK))
-
-	namespaceHeader := resp.Header.Get("x-inference-namespace")
-	podHeader := resp.Header.Get("x-inference-pod")
-
-	// Read and verify the streaming response
-	respBody, err := io.ReadAll(resp.Body)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-
-	ginkgo.By(fmt.Sprintf("Streaming Chat Completion received response length: %d bytes", len(respBody)))
-
-	return namespaceHeader, podHeader
-}
-
-// runCompletionWithCacheThreshold sends a completion request with cache_hit_threshold parameter.
-// This triggers the decode-first optimization in the shared-storage connector.
-// Returns namespace header, pod header, and the finish reason from the response.
-func runCompletionWithCacheThreshold(prompt string, cacheHitThreshold float64, forceCacheThresholdFinishReason bool) (string, string, string) {
-	ginkgo.By(fmt.Sprintf("Sending Completion Request with cache_hit_threshold=%v, forceCacheThreshold=%v", cacheHitThreshold, forceCacheThresholdFinishReason))
-
-	body := fmt.Sprintf(`{"model":"%s","prompt":"%s","max_tokens":10,"cache_hit_threshold":%v}`, simModelName, prompt, cacheHitThreshold)
-	req, err := http.NewRequest(http.MethodPost, fmt.Sprintf("http://localhost:%s/v1/completions", port), strings.NewReader(body))
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	req.Header.Set("Content-Type", "application/json")
-
-	// Add X-Cache-Threshold header to force the simulator to return cache_threshold finish_reason
-	if forceCacheThresholdFinishReason {
-		req.Header.Set("X-Cache-Threshold-Finish-Reason", "true")
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	defer func() {
-		err := resp.Body.Close()
-		gomega.Expect(err).ToNot(gomega.HaveOccurred())
-	}()
-
-	gomega.Expect(resp.StatusCode).Should(gomega.Equal(http.StatusOK))
-
-	namespaceHeader := resp.Header.Get("x-inference-namespace")
-	podHeader := resp.Header.Get("x-inference-pod")
-
-	// Parse response to get finish_reason
-	respBody, err := io.ReadAll(resp.Body)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-
-	// Extract finish_reason from JSON response
-	finishReason := extractFinishReason(string(respBody))
-
-	ginkgo.By(fmt.Sprintf("Completion Response: ns=%s, pod=%s, finish_reason=%s", namespaceHeader, podHeader, finishReason))
-
-	return namespaceHeader, podHeader, finishReason
-}
-
-// runStreamingCompletionWithCacheThreshold sends a streaming completion request with cache_hit_threshold.
-func runStreamingCompletionWithCacheThreshold(prompt string, cacheHitThreshold float64, forceCacheThresholdFinishReason bool) (string, string, string) {
-	ginkgo.By(fmt.Sprintf("Sending Streaming Completion Request with cache_hit_threshold=%v, forceCacheThreshold=%v", cacheHitThreshold, forceCacheThresholdFinishReason))
-
-	body := fmt.Sprintf(`{"model":"%s","prompt":"%s","max_tokens":10,"stream":true,"cache_hit_threshold":%v}`, simModelName, prompt, cacheHitThreshold)
-	req, err := http.NewRequest(http.MethodPost, fmt.Sprintf("http://localhost:%s/v1/completions", port), strings.NewReader(body))
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	req.Header.Set("Content-Type", "application/json")
-
-	if forceCacheThresholdFinishReason {
-		req.Header.Set("X-Cache-Threshold-Finish-Reason", "true")
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	defer func() {
-		err := resp.Body.Close()
-		gomega.Expect(err).ToNot(gomega.HaveOccurred())
-	}()
-
-	gomega.Expect(resp.StatusCode).Should(gomega.Equal(http.StatusOK))
-
-	namespaceHeader := resp.Header.Get("x-inference-namespace")
-	podHeader := resp.Header.Get("x-inference-pod")
-
-	// Read streaming response and extract finish_reason from the last data chunk
-	respBody, err := io.ReadAll(resp.Body)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-
-	finishReason := extractFinishReasonFromStreaming(string(respBody))
-
-	ginkgo.By(fmt.Sprintf("Streaming Completion Response: ns=%s, pod=%s, finish_reason=%s", namespaceHeader, podHeader, finishReason))
-
-	return namespaceHeader, podHeader, finishReason
-}
-
-// extractFinishReason extracts the finish_reason field from a JSON response string.
-func extractFinishReason(jsonStr string) string {
-	// Simple extraction - look for "finish_reason":"value" pattern
-	idx := strings.Index(jsonStr, `"finish_reason":"`)
-	if idx == -1 {
-		// Try with null value
-		if strings.Contains(jsonStr, `"finish_reason":null`) {
-			return "null"
-		}
-		return ""
-	}
-	start := idx + len(`"finish_reason":"`)
-	end := strings.Index(jsonStr[start:], `"`)
-	if end == -1 {
-		return ""
-	}
-	return jsonStr[start : start+end]
-}
-
-// extractFinishReasonFromStreaming extracts the finish_reason from the last SSE data chunk.
-func extractFinishReasonFromStreaming(sseData string) string {
-	// Find the last "finish_reason" that is not null
-	lines := strings.Split(sseData, "\n")
-	lastFinishReason := ""
-	for _, line := range lines {
-		if strings.HasPrefix(line, "data: ") && !strings.Contains(line, "[DONE]") {
-			fr := extractFinishReason(line)
-			if fr != "" && fr != "null" {
-				lastFinishReason = fr
-			}
-		}
-	}
-	return lastFinishReason
-}
-
-// getPrefillRequestCount gets the total request count from a prefill pod's metrics endpoint.
-// This is used to verify whether a request was processed by the prefill pod.
-func getPrefillRequestCount(prefillPodName string) int {
-	ginkgo.By("Getting request count from prefill pod: " + prefillPodName)
-
-	// Use Kubernetes API proxy to access the metrics endpoint
-	output, err := testConfig.KubeCli.CoreV1().RESTClient().
-		Get().
-		Namespace(nsName).
-		Resource("pods").
-		Name(prefillPodName + ":8000").
-		SubResource("proxy").
-		Suffix("metrics").
-		DoRaw(testConfig.Context)
-	if err != nil {
-		ginkgo.By(fmt.Sprintf("Warning: Could not get metrics from prefill pod %s: %v", prefillPodName, err))
-		return -1
-	}
-
-	return parseRequestCountFromMetrics(string(output))
-}
-
-// parseRequestCountFromMetrics extracts the request count from Prometheus metrics output.
-func parseRequestCountFromMetrics(metricsOutput string) int {
-	// Look for vllm:e2e_request_latency_seconds_count{model_name="food-review"} <count>
-	lines := strings.Split(metricsOutput, "\n")
-	for _, line := range lines {
-		if strings.Contains(line, "vllm:e2e_request_latency_seconds_count") &&
-			strings.Contains(line, "food-review") {
-			// Extract the count value after the last space
-			parts := strings.Fields(line)
-			if len(parts) >= 2 {
-				count, err := strconv.Atoi(parts[len(parts)-1])
-				if err == nil {
-					return count
-				}
-			}
-		}
-	}
-	return 0
-}
-
-// Simple EPP configuration for running without P/D
-const simpleConfig = `apiVersion: inference.networking.x-k8s.io/v1alpha1
-kind: EndpointPickerConfig
-plugins:
-- type: prefix-cache-scorer
-  parameters:
-    maxPrefixBlocksToMatch: 256
-    lruCapacityPerServer: 256
-- type: decode-filter
-- type: max-score-picker
-- type: single-profile-handler
-schedulingProfiles:
-- name: default
-  plugins:
-  - pluginRef: decode-filter
-  - pluginRef: max-score-picker
-  - pluginRef: prefix-cache-scorer
-    weight: 2
-`
-
-// EPP configuration for running with P/D
-const pdConfig = `apiVersion: inference.networking.x-k8s.io/v1alpha1
-kind: EndpointPickerConfig
-featureGates:
-- prepareDataPlugins
-plugins:
-- type: prefill-header-handler
-- type: prefix-cache-scorer
-  parameters:
-    blockSizeTokens: 16
-    maxPrefixBlocksToMatch: 256
-    lruCapacityPerServer: 256
-- type: prefill-filter
-- type: decode-filter
-- type: max-score-picker
-- type: prefix-based-pd-decider
-  parameters:
-    nonCachedTokens: 16
-- type: pd-profile-handler
-  parameters:
-    deciderPluginName: prefix-based-pd-decider
-schedulingProfiles:
-- name: prefill
-  plugins:
-  - pluginRef: prefill-filter
-  - pluginRef: max-score-picker
-  - pluginRef: prefix-cache-scorer
-    weight: 2
-- name: decode
-  plugins:
-  - pluginRef: decode-filter
-  - pluginRef: max-score-picker
-  - pluginRef: prefix-cache-scorer
-    weight: 2
-`
-
-// EPP configuration for running with P/D using the unified disagg-profile-handler
-const pdDisaggConfig = `apiVersion: inference.networking.x-k8s.io/v1alpha1
-kind: EndpointPickerConfig
-featureGates:
-- prepareDataPlugins
-plugins:
-- type: disagg-headers-handler
-- type: prefix-cache-scorer
-  parameters:
-    blockSizeTokens: 16
-    maxPrefixBlocksToMatch: 256
-    lruCapacityPerServer: 256
-- type: prefill-filter
-- type: decode-filter
-- type: max-score-picker
-- type: prefix-based-pd-decider
-  parameters:
-    nonCachedTokens: 16
-- type: disagg-profile-handler
-  parameters:
-    deciders:
-      prefill: prefix-based-pd-decider
-schedulingProfiles:
-- name: prefill
-  plugins:
-  - pluginRef: prefill-filter
-  - pluginRef: max-score-picker
-  - pluginRef: prefix-cache-scorer
-    weight: 2
-- name: decode
-  plugins:
-  - pluginRef: decode-filter
-  - pluginRef: max-score-picker
-  - pluginRef: prefix-cache-scorer
-    weight: 2
-`
-
-// EPP configuration for running decode-only using disagg-profile-handler (no prefill, no encode)
-const disaggDecodeOnlyConfig = `apiVersion: inference.networking.x-k8s.io/v1alpha1
-kind: EndpointPickerConfig
-plugins:
-- type: prefix-cache-scorer
-  parameters:
-    hashBlockSize: 10
-    maxPrefixBlocksToMatch: 256
-    lruCapacityPerServer: 256
-- type: decode-filter
-- type: max-score-picker
-- type: disagg-profile-handler
-schedulingProfiles:
-- name: decode
-  plugins:
-  - pluginRef: decode-filter
-  - pluginRef: max-score-picker
-  - pluginRef: prefix-cache-scorer
-    weight: 2
-`
-
-// EPP config for running with precise prefix scoring (i.e. KV events)
-// Uses UDS tokenizer sidecar for tokenization
-const kvConfig = `apiVersion: inference.networking.x-k8s.io/v1alpha1
-kind: EndpointPickerConfig
-plugins:
-- type: precise-prefix-cache-scorer
-  parameters:
-    tokenProcessorConfig:
-      blockSize: 16
-      hashSeed: "42"
-    kvEventsConfig:
-      zmqEndpoint: tcp://0.0.0.0:5557
-    indexerConfig:
-      prefixStoreConfig:
-        blockSize: 16
-      tokenizersPoolConfig:
-        modelName: Qwen/Qwen2.5-1.5B-Instruct
-        uds:
-          socketFile: "/tmp/tokenizer/tokenizer-uds.socket"
-      kvBlockIndexConfig:
-        enableMetrics: false                  # enable kv-block index metrics (prometheus)
-        metricsLoggingInterval: 6000000000    # log kv-block metrics as well (1m in nanoseconds)
-- type: decode-filter
-- type: max-score-picker
-- type: single-profile-handler
-schedulingProfiles:
-- name: default
-  plugins:
-  - pluginRef: decode-filter
-  - pluginRef: max-score-picker
-  - pluginRef: precise-prefix-cache-scorer
-    weight: 10
-`
-
-// EPP config for running with precise prefix scoring and external tokenizer PrepareData plugin.
-// The tokenizer plugin runs in the PrepareData phase (before scoring) and attaches
-// pre-computed token IDs to the request, so the scorer skips internal tokenization.
-const kvExternalTokenizerConfig = `apiVersion: inference.networking.x-k8s.io/v1alpha1
-kind: EndpointPickerConfig
-featureGates:
-- prepareDataPlugins
-plugins:
-- type: tokenizer
-  parameters:
-    modelName: Qwen/Qwen2.5-1.5B-Instruct
-- type: precise-prefix-cache-scorer
-  parameters:
-    tokenProcessorConfig:
-      blockSize: 16
-      hashSeed: "42"
-    kvEventsConfig:
-      zmqEndpoint: tcp://0.0.0.0:5557
-    indexerConfig:
-      prefixStoreConfig:
-        blockSize: 16
-      tokenizersPoolConfig:
-        modelName: Qwen/Qwen2.5-1.5B-Instruct
-        uds:
-          socketFile: "/tmp/tokenizer/tokenizer-uds.socket"
-      kvBlockIndexConfig:
-        enableMetrics: false
-- type: decode-filter
-- type: max-score-picker
-- type: single-profile-handler
-schedulingProfiles:
-- name: default
-  plugins:
-  - pluginRef: decode-filter
-  - pluginRef: max-score-picker
-  - pluginRef: precise-prefix-cache-scorer
-    weight: 10
-`
-
-// EPP configuration for running scale model server test
-const scaleConfig = `apiVersion: inference.networking.x-k8s.io/v1alpha1
-kind: EndpointPickerConfig
-plugins:
-- type: max-score-picker
-- type: single-profile-handler
-schedulingProfiles:
-- name: default
-  plugins:
-  - pluginRef: max-score-picker
-`
-
-// EPP configuration for running with vLLM Data Parallel support
-const dataParallelConfig = `apiVersion: inference.networking.x-k8s.io/v1alpha1
-kind: EndpointPickerConfig
-plugins:
-- type: decode-filter
-- type: max-score-picker
-- type: data-parallel-profile-handler
-schedulingProfiles:
-- name: default
-  plugins:
-  - pluginRef: decode-filter
-  - pluginRef: max-score-picker
-`
